[CHORE] LetterImageView 생성 및 ui component 코드 이동(#381)
